requirements.txt
asgiref==3.2.10
Django==3.1
django-debug-toolbar==2.2
django-model-utils==4.0.0
django-print-sql==2018.3.6
factory-boy==3.2.0
Faker==5.7.0
numpy==1.19.5
pandas==1.2.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
text-unidecode==1.3
python-dotenv == 0.15.0
django-extensions==3.1.0

